Fixed some review comments(renamed packages, moved some classes in packages)
