+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,15 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".activities.NotesListActivity">
+        <activity android:name=".activity.NotesListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".activities.AddNoteActivity"/>
-        <activity android:name=".activities.NoteInfoActivity"/>
+        <activity android:name=".activity.AddNoteActivity"/>
+        <activity android:name=".activity.NoteInfoActivity"/>
     </application>
 
 </manifest>
