+++ app/src/main/java/com/example/todolist/adapter/ListNotesAdapter.kt
@@ -1,15 +1,17 @@
-package com.example.todolist.views
+package com.example.todolist.adapter
 
 import android.support.v7.widget.RecyclerView
-import com.example.todolist.models.Note
+import com.example.todolist.model.Note
 import android.content.Context
 import android.content.Intent
 import android.view.ViewGroup
 import android.view.LayoutInflater
+import android.view.View
+import android.widget.TextView
 import com.example.todolist.R
-import com.example.todolist.enums.ExtraNames
+import com.example.todolist.utils.Constants
 
-class ListNotesAdapter : RecyclerView.Adapter<NoteViewHolder> {
+class ListNotesAdapter : RecyclerView.Adapter<ListNotesAdapter.NoteViewHolder> {
 
     private var notes: List<Note>
     private var inflater: LayoutInflater
@@ -31,12 +33,24 @@ class ListNotesAdapter : RecyclerView.Adapter<NoteViewHolder> {
         var note = notes[position]
         holder.fillFields(note)
         holder.titleView.setOnClickListener{
-            intent.putExtra(ExtraNames.NOTE.Name, note)
+            intent.putExtra(Constants.NOTE_KEY_FOR_NOTE_INFO_ACTIVITY, note)
             inflater.context.startActivity(intent)
         }
     }
 
     override fun getItemCount(): Int {
         return notes.size
+    }
+
+    class NoteViewHolder : RecyclerView.ViewHolder {
+        val titleView: TextView
+
+        constructor(itemView: View): super(itemView){
+            titleView = itemView.findViewById(R.id.titleNote)
+        }
+
+        fun fillFields(note: Note){
+            titleView.text = note.Title
+        }
     }
 }
