--- app/src/main/java/com/example/todolist/enums/ExtraNames.kt
@@ -1,5 +1,0 @@
-package com.example.todolist.enums
-
-enum class ExtraNames(val Name: String) {
-    NOTE("Note")
-}
