+++ app/src/main/java/com/example/todolist/model/Note.kt
@@ -1,10 +1,9 @@
-package com.example.todolist.models
+package com.example.todolist.model
 
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
 import java.io.Serializable
-import java.util.*
 
 @Entity
 data class Note (
