+++ app/src/main/java/com/example/todolist/utils/Constants.kt
@@ -1,0 +1,9 @@
+package com.example.todolist.utils
+
+class Constants {
+    companion object {
+
+        //The names of the extra data
+        const val NOTE_KEY_FOR_NOTE_INFO_ACTIVITY: String = "activity.NoteInfoActivity.Note" // The name of Note object for pass in NoteInfoActivity
+    }
+}
