+++ app/src/main/res/layout/activity_add_note.xml
@@ -3,7 +3,7 @@
               android:layout_width="match_parent"
               android:layout_height="match_parent" xmlns:tools="http://schemas.android.com/tools"
               android:orientation="vertical"
-              tools:context=".activities.AddNoteActivity">
+              tools:context=".activity.AddNoteActivity">
     <EditText
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
