+++ app/src/main/res/layout/activity_note_info.xml
@@ -3,7 +3,7 @@
               android:layout_width="match_parent"
               android:layout_height="match_parent" xmlns:tools="http://schemas.android.com/tools"
               android:orientation="vertical"
-              tools:context=".activities.NoteInfoActivity">
+              tools:context=".activity.NoteInfoActivity">
     <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
