+++ app/src/main/res/layout/activity_notes_list.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".activities.NotesListActivity">
+        tools:context=".activity.NotesListActivity">
     <Button
             android:text="@string/add_btn_name_for_activity_notes_list"
             android:layout_width="match_parent"
