+++ app/src/main/res/layout/note_list_item.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".activities.NotesListActivity">
+        tools:context=".activity.NotesListActivity">
     <android.support.v7.widget.CardView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
